refactor(trello): share concurrency mapper
